Feature/control c (#1)
* Now supports process.SendControlC() and sending controlc before kill

* control+c process throws Win32Exceptions which automatically unpack GetLastError and reports its equivalent user readable message. Observable process now automatically deals with bat files's quit prompt

+++ src/Proc/ObservableProcessBase.cs
@@ -1,8 +1,10 @@
 ﻿using System;
 using System.ComponentModel;
 using System.Diagnostics;
+using System.Globalization;
 using System.IO;
 using System.Reactive.Linq;
+using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using ProcNet.Std;
@@ -19,9 +21,7 @@ namespace ProcNet
 		protected bool NoWrapInThread { get; }
 
 		protected ObservableProcessBase(string binary, params string[] arguments)
-			: this(new StartArguments(binary, arguments))
-		{
-		}
+			: this(new StartArguments(binary, arguments)) { }
 
 		public StreamWriter StandardInput => this.Process.StandardInput;
 
@@ -45,7 +45,7 @@ namespace ProcNet
 
 		public int? ExitCode { get; private set; }
 
-		public int? ProcessId => this.Process?.Id;
+		public virtual int? ProcessId => this.Process?.Id;
 
 		protected IObservable<TConsoleOut> OutStream { get; private set; } = Observable.Empty<TConsoleOut>();
 
@@ -81,7 +81,7 @@ namespace ProcNet
 			}
 			finally
 			{
-				if (!started) this._completedHandle.Set();
+				if (!started) this.SetCompletedHandle();
 			}
 
 			return false;
@@ -90,6 +90,8 @@ namespace ProcNet
 		protected virtual void OnError(IObserver<TConsoleOut> observer, Exception e) => observer.OnError(e);
 
 		protected virtual void OnCompleted(IObserver<TConsoleOut> observer) => observer.OnCompleted();
+
+		private readonly object _exitLock = new object();
 
 		protected void OnExit(IObserver<TConsoleOut> observer)
 		{
@@ -101,10 +103,19 @@ namespace ProcNet
 			}
 			finally
 			{
-				if (!_isDisposing)
-				{
-					this.Stop(exitCode, observer);
-				}
+				ExitStop(observer, exitCode);
+			}
+		}
+
+		private void ExitStop(IObserver<TConsoleOut> observer, int? exitCode)
+		{
+			if (!this.Started) return;
+			if (_isDisposing) return;
+			lock (_exitLock)
+			{
+				if (!this.Started) return;
+
+				this.Stop(exitCode, observer);
 			}
 		}
 
@@ -123,7 +134,8 @@ namespace ProcNet
 				RedirectStandardInput = true
 			};
 			if (s.Environment != null)
-				foreach (var kv in s.Environment) processStartInfo.Environment[kv.Key] = kv.Value;
+				foreach (var kv in s.Environment)
+					processStartInfo.Environment[kv.Key] = kv.Value;
 
 			if (!string.IsNullOrWhiteSpace(s.WorkingDirectory)) processStartInfo.WorkingDirectory = s.WorkingDirectory;
 
@@ -149,16 +161,62 @@ namespace ProcNet
 		{
 			if (WaitHandle.WaitAll(CompletionHandles(), timeout)) return true;
 
-			//if (this._completedHandle.WaitOne(timeout)) return true;
-
 			this.Stop();
 			return false;
 		}
 
+		private readonly object _unpackLock = new object();
+		private readonly object _sendLock = new object();
+		private bool _sentControlC = false;
+
+		public void SendControlC()
+		{
+			if (_sentControlC) return;
+			if (!this.ProcessId.HasValue) return;
+			var path = Path.Combine(Path.GetTempPath(), "proc-c.exe");
+			this.UnpackTempOutOfProcessSignalSender(path);
+			lock (_sendLock)
+			{
+				if (_sentControlC) return;
+				if (!this.ProcessId.HasValue) return;
+				var args = new StartArguments(path, this.ProcessId.Value.ToString(CultureInfo.InvariantCulture))
+				{
+					WaitForExit = null,
+					CallingControlC = true
+				};
+				var result = Proc.Start(args, TimeSpan.FromSeconds(2));
+				_sentControlC = true;
+			}
+		}
+
+		private void UnpackTempOutOfProcessSignalSender(string path)
+		{
+			if (File.Exists(path)) return;
+			var assembly = typeof(Proc).GetTypeInfo().Assembly;
+			try
+			{
+				lock (_unpackLock)
+				{
+					if (File.Exists(path)) return;
+					using (var stream = assembly.GetManifestResourceStream("ProcNet.Embedded.Proc.ControlC.exe"))
+					using (var fs = File.OpenWrite(path))
+						stream.CopyTo(fs);
+				}
+			}
+			catch (Exception e)
+			{
+				Console.WriteLine(e);
+				throw;
+			}
+		}
+
+		protected bool StopRequested => _stopRequested || _sentControlC;
+		private bool _stopRequested;
 		private void Stop(int? exitCode = null, IObserver<TConsoleOut> observer = null)
 		{
 			try
 			{
+				this._stopRequested = true;
 				if (this.Process == null) return;
 
 				var wait = this._startArguments.WaitForExit;
@@ -166,10 +224,22 @@ namespace ProcNet
 				{
 					if (this.Started && wait.HasValue)
 					{
-						this.Process?.Kill();
-						var exitted = this.Process?.WaitForExit((int) wait.Value.TotalMilliseconds);
-						if (this.Process != null && !exitted.GetValueOrDefault(false))
-							this.HardWaitForExit(TimeSpan.FromSeconds(10));
+						bool exitted;
+						if (this._startArguments.SendControlCFirst)
+						{
+							this.SendControlC();
+							exitted = this.Process?.WaitForExit((int) wait.Value.TotalMilliseconds) ?? false;
+							//still attempt to kill to process if control c failed
+							if (!exitted) this.Process?.Kill();
+						}
+						else
+						{
+							this.Process?.Kill();
+							exitted = this.Process?.WaitForExit((int) wait.Value.TotalMilliseconds) ?? false;
+						}
+
+						//if we haven't exited do a hard wait for exit by using the overload that does not timeout.
+						if (this.Process != null && !exitted) this.HardWaitForExit(TimeSpan.FromSeconds(10));
 					}
 					else if (this.Started)
 					{
@@ -201,9 +271,22 @@ namespace ProcNet
 
 				this.Started = false;
 				if (observer != null) OnCompleted(observer);
-				this._completedHandle.Set();
+				this.SetCompletedHandle();
 			}
 		}
+
+		private void SetCompletedHandle()
+		{
+			OnBeforeSetCompletedHandle();
+			this._completedHandle.Set();
+		}
+
+		protected virtual void OnBeforeSetCompletedHandle()
+		{
+
+		}
+
+
 
 		private bool HardWaitForExit(TimeSpan timeSpan)
 		{
